Change arg checking to explicit check for None
When checking for None, it's better to explicitly test

    x = val if val is not None else default

instead of

    x = val if val else default

If the val cannot be implicitly converted to a bool (but is otherwise
valid) the second test will throw an exeption.

As an example of code that this commit fixes, consider the following:

    G = Graph(np.arange(10))

Before this commit, this would fail with the error
"ValueError: The truth value of an array with more than one element is ambiguous."
After this commit, it succeeds as we would expect.